pulled out part/conn dossier and fixed to working
